// generated/src/app/api-models/user-def-attribs-token.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class UserDefAttribsToken {
  winId: string;
  datawindowId: string;
  columnName: string;
  languageId: number;
  userDefinedFieldLabel: string;
  userDefinedDisplayMask: string;
  userDefinedValidLengths: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}